Update project for Xcode 9.3 and fix new build errors.
Fixes #82.

+++ Sources/HTMLParser.m
@@ -2827,7 +2827,7 @@ static BOOL IsHTMLIntegrationPoint(HTMLElement *node)
             }
             
         default:
-            NSAssert(NO, @"cannot handle %@ token in insertion mode %zd", [token class], insertionMode);
+            NSAssert(NO, @"cannot handle %@ token in insertion mode %ld", [token class], (long)insertionMode);
             break;
     }
 }
